+++ tools/bootstrap/bootstrap_algos.cpp
@@ -1,47 +1,49 @@
-#include <iostream>
-#include <vector>
-#include <cstdlib> // For rand and srand functions
-#include <ctime>   // For time function
+#include "bootstrap_algos.h"
 
+using Eigen::VectorXd;
 using namespace std;
 
 // Function to perform block bootstrap resampling
-vector<double> blockBootstrap(const vector<double>& samples, int blockSize) {
+VectorXd blockBootstrap(const VectorXd& samples, int blockSize) {
+    std::random_device rd;
+    std::mt19937 gen(rd());
+    std::uniform_int_distribution<> dis(0, samples.size() - 1);
     // Seed the random number generator with the current time
-    srand(static_cast<unsigned int>(time(0)));
-
-    // Get the size of the original samples
+   // srand(static_cast<unsigned int>(time(0)));
     int originalSize = samples.size();
+    int numBlocks = originalSize / blockSize; // full blocks
+    int remainingSamples = originalSize % blockSize; // To deal with the case of originalSize not a multiple of blockSize
+    VectorXd resampledData(originalSize);
 
-    // Create a vector to store the resampled data
-    vector<double> resampledData;
-
-    // Perform block bootstrap resampling
-    for (int i = 0; i < originalSize; i += blockSize) {
-        // Generate a random index to select a block
+    // Perform block bootstrap resampling for full blocks
+    for (int i = 0; i < numBlocks; ++i) {
+        // Select a block
         int blockIndex = rand() % (originalSize - blockSize + 1);
-
-        // Append the selected block to the resampled data
-        for (int j = 0; j < blockSize; ++j) {
-            resampledData.push_back(samples[blockIndex + j]);
-        }
+        resampledData.segment(i * blockSize, blockSize) = samples.segment(blockIndex, blockSize);
     }
-
+    // Handling the remaining samples
+    if (remainingSamples > 0) {
+        // Select a block
+        int blockIndex = rand() % (originalSize - blockSize + 1);
+        resampledData.tail(remainingSamples) = samples.segment(blockIndex, remainingSamples);
+    }
     return resampledData;
 }
 
 // Function to perform classical bootstrap resampling
-std::vector<double> Bootstrap(const std::vector<double>& samples) {
-    srand(static_cast<unsigned int>(time(0)));
+VectorXd Bootstrap(const VectorXd& samples) {
+    std::random_device rd;
+    std::mt19937 gen(rd());
+    std::uniform_int_distribution<> dis(0, samples.size() - 1);
+    //srand(static_cast<unsigned int>(time(0)));
     int originalSize = samples.size();
-    std::vector<double> resampledData;
+    VectorXd resampledData(originalSize);
 
     // Perform classical bootstrap resampling
     for (int i = 0; i < originalSize; ++i) {
         // Generate a random index to select an element (with replacement)
         int randomIndex = rand() % originalSize;
-        resampledData.push_back(samples[randomIndex]);
+        resampledData[i]=samples[randomIndex];
     }
-
     return resampledData;
 }